requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0

# NOTE: Use newer Pydantic to ensure Python 3.14 wheels exist
pydantic>=2.10.0
pydantic-settings>=2.6.0
google-genai>=0.3.0
python-multipart>=0.0.9
Unidecode>=1.3.8
Metaphone>=0.6
